# feat(perf): support reading symbols from jitdump