## Motivation

The agent already supports reading symbols from [perf map files](https://github.com/torvalds/linux/blob/master/tools/perf/Documentation/jit-interface.txt), and there is a second format called `jitdump`:

https://github.com/torvalds/linux/blob/master/tools/perf/Documentation/jitdump-specification.txt

Some languages with `jitdump` support:

* Elixir/Erlang (#145)
* JVM via perf's [`libperf_jvmti.so` agent library](https://github.com/torvalds/linux/tree/master/tools/perf/jvmti) (#1)
* Node.js, Chrome V8 Javascript
* Wasmtime: https://docs.wasmtime.dev/examples-profiling-perf.html

See also https://developer.arm.com/documentation/101816/0800/JIT-profiling-support/JIT-profiling

## Changes

* Add a parser for the JITDUMP format
	* Everything is parsed, expect the EH Frames (if any) for now 
	* ~~It is currently optimistic and warns if a field cannot be read, hoping the rest can be (happy to discuss if it is good or bad)~~
	* They maybe some refactoring possible regarding how reads are done, I expect a lot of feedback :grinning:
* Add JITDUMP as a fallback to Perf map, looking for the file path in `/proc/<pid>/maps` (the file is `mmap()`'d to be used as marker)

Currently only JIT Code Loads records are used to mimic a Perf map, but the format has potential for more, it could be used to build Debug Info and upload to Parca, but the problem is JIT Code does not have a build ID. Perf Maps keep priority over jitdumps as they are more lightweight, but if we find advantages in the richer format, it might change in the future. (Otherwise, I might reconfigure the parser to discard unnecessary data as an optimization)

## Testing

I ~~have in mind to add~~ added jitdumps and their JSON snapshot to https://github.com/parca-dev/testdata

Right now, you can:

* Run NodeJS with the `--perf-prof` flag
* Run Erlang with the `ERL_FLAGS='+JPperf dump'` environment variable
* Run Wasmtime with the `--jitdump` flag