@maxbrunet 2022-11-30
> I'm happy to merge this. But I see several to do items. Do you plan to handle them in this PR or shall we merge this?

Yes, I'd like to handle them here, but before I do so, I would appreciate some feedback from @v-thakkar and/or @javierhonduco since self-requested a review :pray: