@maxbrunet 2022-12-09
I added tests, here is a first benchmark:

```console
$ ls -l testdata/jitdump/*.dump
-rw-r--r-- 1 maxime users   506419 Dec  9 12:11 testdata/jitdump/dotnet.dump
-rw-r--r-- 1 maxime users 13040830 Dec  9 12:11 testdata/jitdump/erlang.dump
-rw-r--r-- 1 maxime users    10582 Dec  9 12:11 testdata/jitdump/julia.dump
-rw-r--r-- 1 maxime users 13082624 Dec  9 12:11 testdata/jitdump/libperf-jvmti.dump
-rw-r--r-- 1 maxime users  1562469 Dec  9 12:11 testdata/jitdump/nodejs.dump
-rw-r--r-- 1 maxime users     6516 Dec  9 12:11 testdata/jitdump/php.dump
-rw-r--r-- 1 maxime users   129246 Dec  9 12:11 testdata/jitdump/wasmtime.dump

$ go test -bench=. -benchmem ./pkg/jit
goos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/jit
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
BenchmarkLoadJITDump/dotnet-8               1468            969714 ns/op          692432 B/op      14414 allocs/op
BenchmarkLoadJITDump/erlang-8                 34          34686926 ns/op        22501897 B/op     835150 allocs/op
BenchmarkLoadJITDump/julia-8               21128             49061 ns/op           26057 B/op       1147 allocs/op
BenchmarkLoadJITDump/libperf-jvmti-8          39          27826079 ns/op        20026386 B/op     695169 allocs/op
BenchmarkLoadJITDump/nodejs-8                448           2826668 ns/op         2078581 B/op      39746 allocs/op
BenchmarkLoadJITDump/php-8                 38241             28755 ns/op           17012 B/op        538 allocs/op
BenchmarkLoadJITDump/wasmtime-8             7779            217346 ns/op          174055 B/op       3231 allocs/op
PASS
ok      github.com/parca-dev/parca-agent/pkg/jit        11.512s
```