@javierhonduco 2023-02-01
Oh, another minor thing, now we have the latest `testdata` pulled in, but whenever we grep something in the repo (e.g. using ripgrep) it might match the binary files. Could you please add them to the exclusion list for binary / generated files? Cheers!