@maxbrunet 2023-02-02
> Could you please add them to the exclusion list for binary / generated files?

Sorry, what exclusion list are you referring to? I assume that would be in the `testdata` repo