@maxbrunet 2023-02-02
New benchmark: -60% allocs! Thank you so much @javierhonduco @brancz for spotting this!

```
$ go test -bench=. -benchmem ./pkg/jit
goos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/jit
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
BenchmarkLoadJITDump/dotnet-8               1426            752507 ns/op          630455 B/op       4453 allocs/op
BenchmarkLoadJITDump/erlang-8                 44          25866619 ns/op        19224474 B/op     295672 allocs/op
BenchmarkLoadJITDump/julia-8               26508             38855 ns/op           22103 B/op        445 allocs/op
BenchmarkLoadJITDump/libperf-jvmti-8          51          22224994 ns/op        17721051 B/op     272435 allocs/op
BenchmarkLoadJITDump/nodejs-8                504           2339107 ns/op         1906934 B/op      12791 allocs/op
BenchmarkLoadJITDump/php-8                 56421             20795 ns/op           14843 B/op        180 allocs/op
BenchmarkLoadJITDump/wasmtime-8             9472            148831 ns/op          160221 B/op       1010 allocs/op
PASS
ok      github.com/parca-dev/parca-agent/pkg/jit        10.304s
```

<details>
<summary>Benchstat before/after</summary>

```
goos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/jit
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
                            │   old.txt    │               new.txt                │
                            │    sec/op    │    sec/op     vs base                │
LoadJITDump/dotnet-8          1005.9µ ± 6%   690.8µ ± 12%  -31.32% (p=0.000 n=10)
LoadJITDump/erlang-8           34.77m ± 1%   24.12m ±  1%  -30.63% (p=0.000 n=10)
LoadJITDump/julia-8            51.37µ ± 5%   36.74µ ±  3%  -28.48% (p=0.000 n=10)
LoadJITDump/libperf-jvmti-8    28.41m ± 1%   21.14m ±  2%  -25.62% (p=0.000 n=10)
LoadJITDump/nodejs-8           2.990m ± 7%   2.145m ±  9%  -28.26% (p=0.000 n=10)
LoadJITDump/php-8              27.45µ ± 2%   20.86µ ±  2%  -24.00% (p=0.000 n=10)
LoadJITDump/wasmtime-8         204.3µ ± 6%   149.3µ ±  3%  -26.90% (p=0.000 n=10)
geomean                        978.0µ        704.9µ        -27.93%

                            │   old.txt    │               new.txt                │
                            │     B/op     │     B/op      vs base                │
LoadJITDump/dotnet-8          676.2Ki ± 0%   615.7Ki ± 0%   -8.95% (p=0.000 n=10)
LoadJITDump/erlang-8          21.46Mi ± 0%   18.33Mi ± 0%  -14.57% (p=0.000 n=10)
LoadJITDump/julia-8           25.45Ki ± 0%   21.58Ki ± 0%  -15.18% (p=0.000 n=10)
LoadJITDump/libperf-jvmti-8   19.10Mi ± 0%   16.90Mi ± 0%  -11.51% (p=0.000 n=10)
LoadJITDump/nodejs-8          1.982Mi ± 0%   1.819Mi ± 0%   -8.26% (p=0.000 n=10)
LoadJITDump/php-8             16.61Ki ± 0%   14.50Ki ± 0%  -12.75% (p=0.000 n=10)
LoadJITDump/wasmtime-8        170.0Ki ± 0%   156.5Ki ± 0%   -7.95% (p=0.000 n=10)
geomean                       636.6Ki        564.4Ki       -11.35%

                            │   old.txt    │               new.txt               │
                            │  allocs/op   │  allocs/op   vs base                │
LoadJITDump/dotnet-8          14.414k ± 0%   4.453k ± 0%  -69.11% (p=0.000 n=10)
LoadJITDump/erlang-8           835.1k ± 0%   295.7k ± 0%  -64.60% (p=0.000 n=10)
LoadJITDump/julia-8            1147.0 ± 0%    445.0 ± 0%  -61.20% (p=0.000 n=10)
LoadJITDump/libperf-jvmti-8    695.2k ± 0%   272.4k ± 0%  -60.81% (p=0.000 n=10)
LoadJITDump/nodejs-8           39.75k ± 0%   12.79k ± 0%  -67.82% (p=0.000 n=10)
LoadJITDump/php-8               538.0 ± 0%    180.0 ± 0%  -66.54% (p=0.000 n=10)
LoadJITDump/wasmtime-8         3.231k ± 0%   1.010k ± 0%  -68.74% (p=0.000 n=10)
geomean                        18.21k        6.247k       -65.69%
```

</details>